made structuring of files and dirrectories after first homework created files for interfaces and imported them where it needed
