+++ src/components/StartPage/NumberInputContainer/NumberInput/interfaces/NumberInput.interface.ts
@@ -1,0 +1,6 @@
+export interface numberInputProps {
+    placeholder: string,
+    label: string,
+    min: number,
+    max: number
+}
