+++ src/components/StartPage/SelectsContainer/SelectInput/interfaces/SelectInput.interface.ts
@@ -1,0 +1,6 @@
+import { Option } from "../../options/option.interface"
+
+export interface selectInputProps {
+    label: string,
+    options: Option[]
+}
