+++ src/components/StartPage/SelectsContainer/options/options.ts
@@ -1,0 +1,18 @@
+import { Option } from "./option.interface";
+
+const categories: Option[] = [
+    { value: 'example', label: 'Example' },
+];
+const difficulties: Option[] = [
+    { value: 'example', label: 'Example' },
+];
+const types: Option[] = [
+    { value: 'example', label: 'Example' },
+];
+const times: Option[] = [
+    { value: '1m', label: '1m' },
+    { value: '2m', label: '2m' },
+    { value: '5m', label: '5m' },
+];
+
+export { categories, difficulties, types, times };
